pyproject.toml
[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[project]
name = "baizefindb"
version = "0.1.0"
description = "Personal AI financial radar and research assistant."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "akshare>=1.18.59,<2",
    "alembic>=1.14,<2",
    "asyncpg>=0.30,<1",
    "celery>=5.4,<6",
    "fastapi>=0.115,<1",
    "pandas>=2.3,<3",
    "pydantic-settings>=2.7,<3",
    "redis>=5.2,<6",
    "sqlalchemy[asyncio]>=2.0,<3",
    "uvicorn[standard]>=0.34,<1",
]

[dependency-groups]
dev = [
    "httpx>=0.28,<1",
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.25,<1",
    "ruff>=0.9,<1",
]

[tool.setuptools]
package-dir = {"" = "backend"}

[tool.setuptools.packages.find]
where = ["backend"]

[tool.pytest.ini_options]
pythonpath = ["backend"]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.lint.isort]
known-first-party = ["app"]
